email changes commited

+++ event-api/src/main/kotlin/com/und/service/eventapi/EventService.kt
@@ -81,7 +81,7 @@ class EventService {
 
     fun buildEvent(fromEvent: Event, request: HttpServletRequest): Event {
         with(fromEvent) {
-            clientId = tenantProvider.tenant.toInt()
+            clientId = tenantProvider.tenant.toLong()
             ipAddress = request.ipAddr()
             timeZone = AuthenticationUtils.principal.timeZoneId
             agentString = request.getHeader("User-Agent")
